refactor: Controller 클래스 분리 및 함수로 변경
